Fix grouping order in find results
